feat(theme): добавлено управление темой, убран styled-components, добавлен sass для работы

+++ src/shared/store/AuthStore/AuthStore.ts
@@ -4,14 +4,21 @@ import {notification} from 'antd';
 
 import {TImageKeys} from 'shared/assets';
 
+const IMAGE_LIST_NAME: TImageKeys[] = [
+  '1.jpg',
+  '2.jpg',
+  '3.jpg',
+  '4.jpg',
+  '5.jpg',
+];
 export interface IUserData {
   id: number;
-  first_name: string;
+  userNameInfo: string | null;
   avatar: TImageKeys;
 }
 
 function isTImageKey(value: string | null): value is TImageKeys {
-  const validKeys: TImageKeys[] = ['1.jpg', '2.jpg', '3.jpg', '4.jpg', '5.jpg'];
+  const validKeys: TImageKeys[] = IMAGE_LIST_NAME;
   return validKeys.includes(value as TImageKeys);
 }
 
@@ -21,7 +28,10 @@ class AuthStore {
   loading: boolean = false;
   login: string = '';
   password: string = '';
-  availableImages: TImageKeys[] = ['1.jpg', '2.jpg', '3.jpg', '4.jpg', '5.jpg'];
+  userName: string = '';
+  availableImages: TImageKeys[] = IMAGE_LIST_NAME;
+  isLoginInvalid: boolean = false;
+  isPasswordInvalid: boolean = false;
 
   constructor() {
     makeAutoObservable(this);
@@ -29,28 +39,86 @@ class AuthStore {
   }
 
   setLogin = action((value: string) => {
-    this.login = value;
+    this.login = value.toLowerCase().trim();
+    this.validateLogin();
+  });
+
+  resetForm = action(() => {
+    this.login = '';
+    this.password = '';
+    this.userName = '';
   });
 
   setPassword = action((value: string) => {
-    this.password = value;
+    this.password = value.trim();
+    this.validatePassword();
+  });
+
+  setUserName = action((value: string) => {
+    this.userName = value.trimStart();
+  });
+
+  isFormValid = action((): boolean => {
+    return (
+      this.login.trim() !== '' &&
+      this.password.trim() !== '' &&
+      this.userName.trim() !== '' &&
+      !this.isPasswordInvalid
+    );
+  });
+
+  validatePassword = action(() => {
+    if (this.password.length > 5) {
+      this.isPasswordInvalid = false;
+    } else {
+      this.isPasswordInvalid = true;
+    }
+  });
+
+  checkIfLoginExists = action((): boolean => {
+    const storedLogin = localStorage.getItem('signupLogin');
+    return storedLogin === this.login;
+  });
+
+  areFieldsFilled = action((...fields: string[]): boolean => {
+    return fields.every((field) => field.trim() !== '');
+  });
+
+  // функция проверки логина на ввода латиницей
+  validateLogin = action(() => {
+    const latinRegex = /^[A-Za-z]+$/;
+    if (this.login && !latinRegex.test(this.login)) {
+      this.isLoginInvalid = true;
+    } else {
+      this.isLoginInvalid = false;
+    }
   });
 
   // я решил затащить логику присваивания аватарок пользователю, при авторизации пользователю присваивается рандомная аватарка
-
   private getRandomImage = (): TImageKeys => {
     const randomIndex = Math.floor(Math.random() * this.availableImages.length);
     return this.availableImages[randomIndex];
   };
 
   signUp = action((onSuccess: () => void) => {
+    if (this.checkIfLoginExists()) {
+      notification.error({
+        message: 'Ошибка',
+        description:
+          'Пользователь с таким логином уже существует. Введите другой логин.',
+      });
+      return;
+    }
+
     this.loading = true;
 
     setTimeout(() => {
       // я использую шифрование пароля, не безопасно вообще, но лучше чем ничего
+      const hashedLogin = SHA256(this.login).toString();
       const hashedPassword = SHA256(this.password).toString();
-      // так делать категорически нельзя, но мне хотелось сделать авторизацию
-      localStorage.setItem('signupLogin', this.login);
+      // так делать категорически нельзя!!! Сделано для задания
+      localStorage.setItem('signupLogin', hashedLogin);
+      localStorage.setItem('signupUserName', this.userName);
       localStorage.setItem('signupPassword', hashedPassword);
 
       const randomImage = this.getRandomImage();
@@ -62,6 +130,7 @@ class AuthStore {
       });
 
       this.loading = false;
+      this.resetForm();
       onSuccess();
     }, 2000); // по заданию сказано симулировать запросы
   });
@@ -70,14 +139,17 @@ class AuthStore {
     this.loading = true;
 
     setTimeout(() => {
-      // так делать категорически нельзя, но мне хотелось сделать авторизацию
-      const storedLogin = localStorage.getItem('signupLogin');
-      const storedPassword = localStorage.getItem('signupPassword');
-      // поэтому я использую шифрование пароля, не безопасно вообще, но лучше чем ничего
+      //  я использую шифрование пароля, не безопасно вообще, но лучше чем ничего
+      const hashedLogin = SHA256(this.login).toString();
       const hashedPassword = SHA256(this.password).toString();
 
+      // так делать категорически нельзя!!! Сделано для задания
+      const storedLogin = localStorage.getItem('signupLogin');
+      const storedPassword = localStorage.getItem('signupPassword');
+      const storedUserName = localStorage.getItem('signupUserName');
+
       if (storedLogin && storedPassword) {
-        if (storedLogin === this.login && storedPassword === hashedPassword) {
+        if (storedLogin === hashedLogin && storedPassword === hashedPassword) {
           let userImage = localStorage.getItem('userImage');
 
           if (!isTImageKey(userImage)) {
@@ -87,7 +159,7 @@ class AuthStore {
 
           this.user = {
             id: 1,
-            first_name: this.login,
+            userNameInfo: storedUserName,
             avatar: userImage as TImageKeys,
           };
           this.isAuthenticated = true;
@@ -114,6 +186,11 @@ class AuthStore {
   });
 
   logout = action(() => {
+    this.isAuthenticated = false;
+    localStorage.removeItem('authUser');
+  });
+
+  onDeleteAccount = action(() => {
     this.user = null;
     this.isAuthenticated = false;
     localStorage.removeItem('authUser');
